requirements.txt
asgiref==3.7.2
Django==4.2.10
django-prometheus==2.3.1
djangorestframework==3.14.0
prometheus-client==0.20.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.9.0
